requirements.txt
streamlit
yfinance
numpy
pandas
pandas_ta
plotly
scikit-learn
tensorflow-cpu